Server - Client connection + GetUser from DB
The client can connect to the Socket Server, they can communicate using JSON. Right now you need the database to make it work, or you can comment out a shitton of code and just do the handshake.
Also the client can ask for a User from the DB, but the server can't deal with invalid info yet.